requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
banks==2.4.1
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.135.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.2.0
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
llama-index==0.14.18
llama-index-cli==0.5.6
llama-index-core==0.14.18
llama-index-embeddings-openai==0.6.0
llama-index-indices-managed-llama-cloud==0.11.0
llama-index-instrumentation==0.5.0
llama-index-llms-openai==0.7.3
llama-index-readers-llama-parse==0.6.0
llama-index-workflows==2.17.1
llama-parse==0.5.20
llama_cloud==1.6.0
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.3
numpy==2.4.3
openai==2.29.0
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
portalocker==3.2.0
propcache==0.4.1
protobuf==7.34.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.2
PyYAML==6.0.3
qdrant-client==1.17.1
regex==2026.2.28
requests==2.32.5
setuptools==82.0.1
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tinytag==2.2.1
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
Wikipedia-API==0.11.0
wrapt==2.1.2
yarl==1.23.0
